image-processing/requirements.txt
Flask==2.0.1
pytest==5.3.0; python_version > "3.0"
pytest==4.6.6; python_version < "3.0"
gunicorn==20.1.0
google-cloud-vision==2.4.2
google-cloud-storage==1.42.2
Wand==0.6.7
pymongo[tls,srv,gssapi]==3.6.1
